added basic test scripts Runs a simple test whether the output of utftex is different from a reference output created at some time in the past (i.e. test for regressions... if you are lucky)
two test scripts are provided:
1: runtest.sh calls utftex on a bunch of test input files in the
test directory
2: testfont.sh uses utftex -t to get a table of available fonts to
see if that still works

Call the scripts with the argument "ref" to create a reference.
Without argument a test is run.

The test files for runtest.sh are collected in the test directory.
If new tests are added one needs to recreate the reference. To get
started I stole some test files from here:
https://github.com/larseggert/asciiTeX/tree/main/examples